backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
python-pptx==0.6.21
pillow==10.1.0
aiofiles==23.2.1
requests==2.31.0
openai==1.3.8
python-pptx==0.6.21
pillow==10.1.0
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2